sigopt factory
