+++ sigopt/experiment_context.py
@@ -6,7 +6,7 @@ from .defaults import assert_valid_project_id, get_default_project, ensure_proje
 
 
 class ExperimentContext(BaseRunFactory):
-  '''Inherits the create_run method from BaseRunFactory.'''
+  '''Wraps the Experiment object and provides extra utility methods.'''
 
   def __init__(self, experiment):
     assert experiment.project, "The experiment must belong to a project"
@@ -63,20 +63,3 @@ class ExperimentContext(BaseRunFactory):
       )
     )
     return run_context
-
-def create_experiment(name, parameters, metrics=None, project=None, budget=None, **kwargs):
-  if project is None:
-    project = get_default_project()
-  else:
-    assert_valid_project_id(project)
-  connection = get_connection()
-  ensure_project_exists(connection, project)
-  experiment = connection.experiments().create(
-    name=name,
-    metrics=metrics,
-    parameters=parameters,
-    project=project,
-    observation_budget=budget,
-    **kwargs,
-  )
-  return ExperimentContext(experiment)
